+++ modules/app_engine/main.tf
@@ -14,19 +14,10 @@
  * limitations under the License.
  */
 
-variable "admin_email" {
-  description = "Admin user email on Gsuite"
-}
-
-variable "organization_id" {
-  description = "The organization id for the associated services"
-}
-
-variable "billing_account" {
-  description = "The ID of the billing account to associate this project with"
-}
-
-variable "credentials_path" {
-  description = "Path to a service account credentials file with rights to run the Project Factory. If this file is absent Terraform will fall back to Application Default Credentials."
-  default     = ""
+resource "google_app_engine_application" "main" {
+  project          = "${var.project_id}"
+  location_id      = "${var.location_id}"
+  auth_domain      = "${var.auth_domain}"
+  serving_status   = "${var.serving_status}"
+  feature_settings = "${var.feature_settings}"
 }