+++ modules/core_project_factory/variables.tf
@@ -83,7 +83,8 @@ variable "usage_bucket_prefix" {
 }
 
 variable "credentials_path" {
-  description = "Path to a Service Account credentials file with permissions documented in the readme"
+  description = "Path to a service account credentials file with rights to run the Project Factory. If this file is absent Terraform will fall back to Application Default Credentials."
+  default     = ""
 }
 
 variable "shared_vpc_subnets" {
@@ -111,12 +112,6 @@ variable "bucket_name" {
 variable "auto_create_network" {
   description = "Create the default network"
   default     = "false"
-}
-
-variable "app_engine" {
-  description = "A map for app engine configuration"
-  type        = "map"
-  default     = {}
 }
 
 variable "disable_services_on_destroy" {