+++ test/integration/full/controls/app-engine.rb
@@ -15,11 +15,6 @@
 domain           = attribute('domain')
 project_id       = attribute('project_id')
 region           = attribute('region')
-credentials_path = attribute('credentials_path')
-
-ENV['CLOUDSDK_AUTH_CREDENTIAL_FILE_OVERRIDE'] = File.absolute_path(
-  credentials_path,
-  File.join(__dir__, "../../../fixtures/full"))
 
 control 'project-factory-app-engine' do
   title "Project Factory App Engine configuration"
@@ -37,7 +32,7 @@ control 'project-factory-app-engine' do
     end
 
     it { expect(metadata).to include(authDomain: domain) }
-    it { expect(metadata).to include(featureSettings: Hash.new) }
+    it { expect(metadata).to include(featureSettings: {:splitHealthChecks=>true}) }
     it { expect(metadata).to include(id: project_id) }
     it { expect(metadata).to include(name: "apps/#{project_id}") }
     it { expect(metadata).to include(locationId: region) }