+++ test/integration/full/controls/shared-vpc.rb
@@ -18,11 +18,6 @@ project_id                  = attribute('project_id')
 project_number              = attribute('project_number')
 service_account_email       = attribute('service_account_email')
 shared_vpc                  = attribute('shared_vpc')
-credentials_path            = attribute('credentials_path')
-
-ENV['CLOUDSDK_AUTH_CREDENTIAL_FILE_OVERRIDE'] = File.absolute_path(
-  credentials_path,
-  File.join(__dir__, "../../../fixtures/full"))
 
 control 'project-factory-shared-vpc' do
   title "Project Factory shared VPC"